requirements.txt
flask
flask-sqlalchemy
flask-socketio
Flask-Login
psycopg2-binary
requests
pydantic
